Draw checkbox using lines.
Fix broken page-based attributes.
Fix incorrectly positioned and sized image on example page.

Checkboxes were previously drawn using a textbox containing "X". When the field is too small, the "X" doesn't even render, resulting in bugs that can only be found manually. With this change, the checkbox is guaranteed to always be drawn.

Closes https://github.com/Gusto/prawn-fillform/pull/5